# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Bible API
  description: |
    Complete Bible API with Role-Based Access Control (RBAC), multi-version support, RSS feeds, user management, file locations, and backup system.
    
    ## Role-Based Access Control
    - **Admin**: Full system access with all permissions
    - **User**: Standard user with Bible reading and format access
    - **Viewer**: Read-only access to Bible content
    - **Editor**: Can edit content but not manage users
    
    ## Authentication
    Use Bearer token authentication. Login to get your JWT token.
    
    ## Rate Limiting
    - General API: 100 requests per 15 minutes
    - Auth endpoints: 5 attempts per 15 minutes
    
    ## File Locations
    Users can configure local drive paths where TXT and XML files will be saved automatically when accessing verses.
  version: 3.0.0
  contact:
    name: API Support
    email: support@bibleapi.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.bibleapi.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /user/location/txt:
    post:
      summary: Set TXT file location
      description: Set the local drive path where TXT files will be saved
      tags: [File Locations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_path
              properties:
                file_path:
                  type: string
                  description: Full local file path including filename
                  example: "D:/BibleFiles/daily_verse.txt"
      responses:
        200:
          description: TXT location set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
        400:
          description: Invalid file path or directory not writable
        401:
          description: Authentication required
        403:
          description: Permission denied
  /user/location/xml:
    post:
      summary: Set XML file location
      description: Set the local drive path where XML files will be saved
      tags: [File Locations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_path
              properties:
                file_path:
                  type: string
                  description: Full local file path including filename
                  example: "D:/BibleFiles/bible_show.xml"
      responses:
        200:
          description: XML location set successfully
        400:
          description: Invalid file path or directory not writable
        401:
          description: Authentication required
        403:
          description: Permission denied
  /user/location:
    get:
      summary: Get both file locations
      description: Returns both TXT and XML file locations for the current user
      tags: [File Locations]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Locations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      txt:
                        type: object
                        nullable: true
                      xml:
                        type: object
                        nullable: true
        401:
          description: Authentication required
  /user/location/{type}:
    get:
      summary: Get specific file location
      description: Returns either TXT or XML file location
      tags: [File Locations]
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [txt, xml]
          description: Location type (txt or xml)
      responses:
        200:
          description: Location retrieved successfully
        401:
          description: Authentication required
        404:
          description: Location not found

    delete:
      summary: Delete file location
      description: Remove file location configuration
      tags: [File Locations]
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [txt, xml]
      responses:
        200:
          description: Location deleted successfully
        401:
          description: Authentication required
...
  /rss/{userId}/{version}/verse:
    get:
      summary: Get RSS verse feed
      description: Returns RSS feed of recent verses accessed by the user
      tags: [RSS Feeds]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          example: admin
        - name: version
          in: path
          required: true
          schema:
            type: string
            enum: [kjv, tam]
          example: kjv
      responses:
        200:
          description: RSS feed
          content:
            application/rss+xml:
              schema:
                type: string
        403:
          description: Format access denied
        404:
          description: User or version not found
  /rss/{userId}/{version}/reference:
    get:
      summary: Get RSS reference feed
      description: Returns RSS feed of recent Bible references accessed by the user
      tags: [RSS Feeds]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: RSS feed
          content:
            application/rss+xml:
              schema:
                type: string
        403:
          description: Format access denied
  /rss/feeds:
    get:
      summary: List RSS feeds for user
      description: Returns list of available RSS feeds for the authenticated user
      tags: [RSS Feeds]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: string
          description: User ID (optional, uses authenticated user if not provided)
      responses:
        200:
          description: List of RSS feeds
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user_id:
                    type: string
                  total_feeds:
                    type: integer
                  feeds:
                    type: array

  # ==================== FORMATS ====================
...
  /admin/location/stats:
    get:
      summary: Get file location statistics
      description: Get statistics about user file locations (Admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Location statistics
        403:
          description: Admin access required

  # ==================== USER MANAGEMENT (ADMIN ONLY) ====================
...